got rid of the scat_it, since it had redundant information; bug fixes & a new macro - center and corresponding atom context menu item are added, this allows to centre the rotation on any particular atom or centre of selection; a Ctrl+Shift + Left mouse button are now association with XY movement of the whole scene; a restriction of maximum of 12 member rings is added to select rings to avoid problems with polymeric structures...

+++ branches/dev/xlib/unitcell.cpp
@@ -793,7 +793,7 @@ void TUnitCell::BuildStructureMap( TArray3D<short>& map, double delta, short val
   // precalculate the sphere/ellipsoid etc coordinates for all distinct scatterers
   const TAsymmUnit& au = GetLattice().GetAsymmUnit();
 
-  TPSTypeList<int, double > scatterers;
+  TPSTypeList<short, double > scatterers;
   for( size_t i=0; i < au.AtomCount(); i++ )  {
     if( au.GetAtom(i).IsDeleted() )  continue;
     size_t ind = scatterers.IndexOfComparable(au.GetAtom(i).GetType().index);
